% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archaeology.R
\name{arch_loglike}
\alias{arch_loglike}
\title{Multiple binomial model likelihood}
\usage{
arch_loglike(param, y)
}
\arguments{
\item{param}{Either a vector with two elements, N and phi, or
a data.frame with two columns, named N and phi}

\item{y}{a data vector}
}
\value{
The log-likelihood for each N,phi combination. The implementation
internally uses the log-gamma function to provide a differentiable
function of N, which allows optim() to treat N as a continuous variable.
If N < max(y), the log-likelihood is defined to be -Inf.
}
\description{
Compute the log-likelihood for multiple Binom(N,phi) observations
}
\examples{
arch_loglike(
  param = data.frame(N = c(5, 10, 20, 40),
                     phi = c(0.1, 0.2, 0.3, 0.4)),
  y = c(2, 4))
}
