% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{filament1_estimate}
\alias{filament1_estimate}
\title{Estimate filament models}
\usage{
filament1_estimate(data, model)
}
\arguments{
\item{data}{A \code{data.frame} with the same variables as the \code{filament1} data set.
Must have columns \code{CAD_Weight} and \code{Actual_Weight}}

\item{model}{Either "A" for a log-linear variance model, or "B" for a proportional
scaling error model}
}
\value{
An estimation object suitable for use with \code{\link[=filament1_predict]{filament1_predict()}}
}
\description{
Estimate filament models with different variance structure
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
