% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{filament1_predict}
\alias{filament1_predict}
\title{Predict values for a filament model}
\usage{
filament1_predict(object, newdata, alpha = 0.05, df = Inf, ...)
}
\arguments{
\item{object}{The estimation object, like the output from \code{\link[=filament1_estimate]{filament1_estimate()}}}

\item{newdata}{The data to use for prediction.
Must have a column \code{CAD_Weight}}

\item{alpha}{The target coverage error for the prediction intervals, Default: 0.05}

\item{df}{The degrees of freedom for t-quantiles for the prediction interval
construction, Default: Inf}

\item{...}{Additional arguments, currently ignored}
}
\value{
A data.frame with variables \code{mean}, \code{sd}, \code{lwr}, and \code{upr}, summarising
the prediction distribution for each row of \code{newdata}.
}
\description{
Uses an estimated filament model to predict new observations
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
