% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompLab-package.R
\docType{package}
\name{StatCompLab-package}
\alias{StatCompLab}
\alias{StatCompLab-package}
\title{StatCompLab: Workshop Material For UoE Statistical Computing}
\description{
Contains Workshop assignments (as vignettes) and learnr tutorials For MATH10093, Statistical Computing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://finnlindgren.github.io/StatCompLab}
  \item \url{https://github.com/finnlindgren/StatCompLab}
  \item Report bugs at \url{https://github.com/finnlindgren/StatCompLab/issues}
}

}
\author{
\strong{Maintainer}: Finn Lindgren \email{finn.lindgren@ed.ac.uk}

}
\keyword{internal}
